+++ lib/handlers/listing.js
@@ -1,3 +1,6 @@
+// Service Wrappers
+var geocoder = require('../wrappers/geocoder.js');
+
 var listing_model_path = '../models/listings/';
 function pt(f) {
     return listing_model_path + f;
@@ -12,6 +15,20 @@ var listing_name_map = {};
 listing_objs.map(function(e, i) {
     listing_name_map[e.name] = e;
 });
+
+function getListingSchema(name) {
+    return listing_objs[name].schema;
+}
+
+function getListingByIDAndUserID(listing_id, user_id, callback) {
+    listing.model.findOne({_id: listing_id, user_id: user_id}, callback);
+}
+exports.getListingsByIDAndUserID = getListingsByIDAndUserID;
+
+function getListingsByUserID(user_id, callback) {
+    listing.model.find({user_id: user_id});
+}
+exports.getListingsByUserID = getListingsByUserID;
 
 function makeNewListingFromPost(body, user_id, callback) {
     body.user_id = user_id;
@@ -112,4 +129,33 @@ function updateSubletListingLatLng(listing, callback) {
         });
     }
 }
+
+function updateMiscListingLatLng(listing, callback) {
+    if (listing.kind != Listings.MISC) {
+        return callback(false);
+    } else if (typeof(listing.address) === typeof(void 0) || typeof(listing.address) != typeof('str') || listing.address.length == 0) {
+        return callback(true);
+    } else {
+        geocoder.geocode(listing.address, function(err, res) {
+            if (!Array.isArray(res)) {
+                return callback(false);
+            }
+            switch(res.length) {
+                case 0:
+                    return callback(false);
+                case 1:
+                    var lat = res[0].latitude;
+                    var lng = res[0].longitude;
+                    listing.apt_info.lat = lat;
+                    listing.apt_info.lng = lng;
+                    return callback(true);
+                default:
+                    console.log("Unexpected number of geocode results for "+new_address);
+                    console.log(res);
+                    return callback(false);
+            }
+        });
+    }
+}
+
