+++ lib/route/listing.js
@@ -1,4 +1,4 @@
-// load Schema
+// Handlers
 var listingHandler = require('../handlers/listing.js');
 
 // Tusk Libraries
@@ -11,48 +11,7 @@ var User = require('../models/user.js')
 var email = require('../email.js')
 
 
-var geocoderProvider = 'google';
-var httpAdapter = 'https';
-// optional
-var extra = {
-    apiKey: 'AIzaSyDE8UDnsYQftCEkjt_k1WuooAez9BsPWCM', // for Mapquest, OpenCage, Google Premier
-    formatter: null         // 'gpx', 'string', ...
-};
-
-var geocoder = require('node-geocoder')(geocoderProvider, httpAdapter, extra);
-
 // Listing Endpoints
-function updateMiscListingLatLng(listing, callback) {
-    if (listing.kind != Listings.MISC) {
-        return callback(false);
-    } else if (typeof(listing.address) === typeof(void 0) || typeof(listing.address) != typeof('str') || listing.address.length == 0) {
-        return callback(true);
-    } else {
-        geocoder.geocode(listing.address, function(err, res) {
-            if (!Array.isArray(res)) {
-                return callback(false);
-            }
-            switch(res.length) {
-                case 0:
-                    return callback(false);
-                case 1:
-                    var lat = res[0].latitude;
-                    var lng = res[0].longitude;
-                    listing.apt_info.lat = lat;
-                    listing.apt_info.lng = lng;
-                    return callback(true);
-                default:
-                    console.log("Unexpected number of geocode results for "+new_address);
-                    console.log(res);
-                    return callback(false);
-            }
-        });
-    }
-}
-
-
-
-
 exports.postListing = function(request, response) {
     response.set('Content-Type', 'application/json');
     if (!Common.ensureLoginSession(request)) {
@@ -61,7 +20,7 @@ exports.postListing = function(request, response) {
             'Listing could not be posted because the user is not logged in'
         ));
     }
-    makeNewListingFromPost(request.body, request.session.login.who.id, function(newListing) {
+    listingHandler.makeNewListingFromPost(request.body, request.session.login.who.id, function(newListing) {
         if (!newListing) {
             return response.status(400).send(Common.error(
                 'LISTING_INVALID_DATA_EXCEPTION',
@@ -96,7 +55,7 @@ exports.postListingById = function(request, response) {
     }
 
     var listing_id = request.params.listing_id;
-    Listings.Listing.findOne({_id: listing_id, user_id: request.session.login.who.id}, function(err, listing) {
+    listingHandler.findListingByIDAndUserID(listing_id, request.session.login.who.id, function(err, listing) {
         if (err || !listing) {
             return response.sendStatus(404);
         } else {
@@ -107,18 +66,18 @@ exports.postListingById = function(request, response) {
                 ));
             }
             for (var key in request.body) {
-                if (!(key in Listings[request.body.type].schema.paths)) {
+                if (!(key in listingHandler.getListingSchema(request.body.type).paths)) {
                     return response.status(400).send(Common.error(
                         'INVALID_LISTING_PATH_FAILURE',
-                        'Listing to be updated did not carry a listing type'
+                        'Listing to be updated contained invalid data'
                     ));
                 }
                 Utils.setValueByPath(listing, key, request.body[key]);
             }
             if ('apt_info.address' in request.body) {
-                updateSubletListingLatLng(listing, function(success) {
+                listingHandler.updateSubletListingLatLng(listing, function(success) {
                     if (success) {
-                        listing.listing.save(function(err) {
+                        listing.save(function(err) {
                             if (err) return response.status(400).send(Common.error(
                                 'LISTING_DATA_VALIDATION_FAILURE_EXCEPTION',
                                 'Listing data failed to validate'
@@ -174,10 +133,7 @@ function secureListing(request, listing, include_id) {
 exports.getListingByUserId = function(request, response) {
     response.set('Content-Type', 'application/json');
 
-    Listings.Listing
-        .find({user_id: request.params.user_id})
-        .limit(Consts.MAX_LISTING_RESULTS)
-        .exec(function(err, listings) {
+    listingHandler.getListingByUserID(request.params.user_id, function(err, listings) {
         if (err) {
             return response.status(400).send(Common.error('Invalid listing by user id request'));
         } else {
